Fortress, medieval and Ancient athens fixes and Jungle Treehouse tweaks (part 3) in 1.20.1 (#10415)
Followup to Medieval and Fortress fixes and Shire and Fortress rebalancing in 1.20.1 #10364.
In the fix of the decorations not being shown as levellable in the build tool, it turned out that the decoration controller broke,
as I forgot to set a pack in the fixed decorations. I thought including a path and schematicName would be enough to tell the decoration controller to not use the legacy code anymore, but it was looking at the pack NBT-component instead.
I've been thinking about removing the legacy code in TileEntityDecorationController as well,
as the decorations for which that was still needed should now all be migrated. But I wasn't sure whether that would be better in a separate PR
Edit: I didn't check whether it would be needed in Stylecolonies, as a few styles there (at least Frontier) stems from before 1.19
Move the medium_military harbour pier decoration in Ancient athens to the correct folder
Rebalance some placeholders on the Jungle Treehouse sifter's hut. Some areas were under-placeholdered, and others were over-placeholdered